requirements.txt
PySide6>=6.6
python-dotenv>=1.0.1
